pyproject.toml
[tool.poetry]
name = "shoulder"
version = "1.2.7"
description = "patient specific anatomic coordinate system generation for shoulder bones"
authors = ["Gregory W Spangenberg <gspangen@westerneng.ca>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/gregspangenberg/shoulder"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
trimesh = "^3.14.1"
scipy = "^1.9.1"
networkx = "^2.8.6"
shapely = "^1.8.4"
rtree = "^1.0.0"
plotly = "^5.10.0"
scikit-spatial = "^6.5.0"
numpy-stl = "^2.17.1"
circle-fit = "^0.1.3"
ruptures = "^1.1.7"
onnxruntime = "1.15.*"
scikit-learn = "^1.2.2"



[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
ipykernel = "^6.15.2"
nbformat = "^5.7.0"
pwlf = "^2.2.1"
pyglet = "~1.5"
ipywidgets = "^7"
pandas = "^2.0.1"
statsmodels = "^0.14.0"
snakeviz = "^2.2.0"
imbalanced-learn = "0.11.*"
xgboost = "^1.7.5"
natsort = "^8.4.0"
dtreeviz = "^2.2.1"
shap = "^0.42.1"
lime = "^0.2.0.1"
skl2onnx = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
